Rename algebra.http to stack.http
