+++ modules/apis/user/src/main/scala/phms/rest/user/UserLoginRoutes.scala
@@ -2,7 +2,7 @@ package phms.rest.user
 
 import org.http4s.dsl._
 import org.http4s._
-import phms.algebra.http._
+import phms.stack.http._
 import phms.algebra.user._
 import phms._
 import phms.kernel._
