+++ modules/apis/user/src/main/scala/phms/rest/user/UserRoutes.scala
@@ -1,7 +1,7 @@
 package phms.rest.user
 
 import org.http4s.dsl._
-import phms.algebra.http._
+import phms.stack.http._
 import phms.algebra.user._
 import phms._
